Grind main/partyobj onDrawDebug: 89.15% -> 89.42%
Cast CFlatCenterState() to int in the early-return guard so the != 0 test
emits a signed cmpwi (matching target) instead of unsigned cmplwi.

+++ src/partyobj.cpp
@@ -5301,7 +5301,7 @@ void CGPartyObj::onDrawDebug(CFont* font, float x, float& y, float z)
 	CGCharaObj::onDrawDebug(font, x, y, z);
 
 	if ((static_cast<signed char>(static_cast<int>((static_cast<unsigned int>(static_cast<unsigned char>(m_weaponNodeFlags)) << 24) & 0xC0000000) >> 31) == 0) ||
-	    (CFlatCenterState() != 0) ||
+	    (static_cast<int>(CFlatCenterState()) != 0) ||
 	    ((MiniGamePcs.m_flags & 0x80) == 0)) {
 		return;
 	}
